refactor(util): Changed the util module to 'struct' to better reflect function
