+++ src/malwoden.ts
@@ -6,6 +6,6 @@ export * as Pathfinding from "./pathfinding";
 export * as Calc from "./calc";
 export * as GUI from "./gui";
 export * as Rand from "./rand";
-export * as Util from "./util";
+export * as Struct from "./struct";
 export * as FOV from "./fov";
-export { Vector2, Rect } from "./util";
+export { Vector2, Rect } from "./struct";
