+++ src/struct/line.spec.ts
