+++ src/struct/rect.spec.ts
