+++ src/struct/rect.ts
