+++ src/struct/table.spec.ts
