+++ src/struct/vector.ts
